adding debugdraw feature and JMath to handle simple 2D graphs like circle and rect.

+++ mario/src/main/java/jade/MouseListener.java
@@ -1,6 +1,7 @@
 package jade;
 
 import org.joml.Vector4d;
+import org.joml.Vector4f;
 
 import static org.lwjgl.glfw.GLFW.GLFW_PRESS;
 import static org.lwjgl.glfw.GLFW.GLFW_RELEASE;
@@ -70,19 +71,19 @@ public class MouseListener {
     return get().yPos;
   }
 
-  public static double getOrthoX() {
-    double currentX = getxPos();
+  public static float getOrthoX() {
+    float currentX = (float) getxPos();
     currentX = (currentX / Window.getWidth()) * 2 - 1;
-    Vector4d tmp = new Vector4d(currentX, 0, 0, 1);
+    Vector4f tmp = new Vector4f(currentX, 0, 0, 1);
     tmp.mul(Window.getCurrentScene().getCamera().getInverseProjectionMatrix()).mul(Window.getCurrentScene().getCamera().getInverseViewMatrix());
     currentX = tmp.x;
     return currentX;
   }
 
-  public static double getOrthoY() {
-    double currentY = Window.getHeight() - getyPos();
+  public static float getOrthoY() {
+    float currentY = (float) Window.getHeight() - (float) getyPos();
     currentY = (currentY / Window.getHeight()) * 2 - 1;
-    Vector4d tmp = new Vector4d(0, currentY, 0, 1);
+    Vector4f tmp = new Vector4f(0, currentY, 0, 1);
     tmp.mul(Window.getCurrentScene().getCamera().getInverseProjectionMatrix()).mul(Window.getCurrentScene().getCamera().getInverseViewMatrix());
     currentY = tmp.y;
     return currentY;
