+++ mario/src/main/java/jade/Prefabs.java
@@ -3,12 +3,13 @@ package jade;
 import components.Sprite;
 import components.SpriteRender;
 import org.joml.Vector2d;
+import org.joml.Vector2f;
 
 public class Prefabs {
 
   public static GameObject generateSpriteWithinGameObject(Sprite sprite, float sizeX, float sizeY) {
     GameObject block = new GameObject("Sprite_Object_GameObject_Generation",
-            new Transform(new Vector2d(), new Vector2d(sizeX, sizeY)), 0);
+            new Transform(new Vector2f(), new Vector2f(sizeX, sizeY)), 0);
     SpriteRender render = new SpriteRender();
     render.setSprite(sprite);
     block.addComponent(render);
